Update hash crate.
