+++ crates/core/hasher/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 moon_error = { path = "../error" }
 moon_task = { path = "../task" }
 moon_utils = { path = "../utils" }
+erased-serde = { workspace = true }
 rustc-hash = { workspace = true }
 serde = { workspace = true }
 sha2 = "0.10.6"
