+++ crates/core/hasher/src/lib.rs
@@ -1,6 +1,10 @@
 mod hasher;
+mod hashset;
 mod helpers;
+mod target_hasher;
 
-pub use hasher::TargetHasher;
+pub use hasher::*;
+pub use hashset::*;
 pub use helpers::*;
 pub use sha2::{Digest, Sha256};
+pub use target_hasher::TargetHasher;
